backend/requirements.txt
# AI Planner Backend

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0

# OpenAI
openai>=1.6.0

# HTTP Client (for calendar APIs)
httpx>=0.26.0

# Redis (optional, for caching)
redis>=5.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
gunicorn>=21.2.0
